requirements.txt
scikit-learn>=1.3.0
skl2onnx>=1.16.0
onnxruntime>=1.16.0
onnx>=1.15.0
mlflow>=2.9.0
wandb>=0.16.0
pandas>=2.1.0
numpy>=1.24.0

flask>=2.2.0,<3.0.0
pydantic>=2.5.0
gunicorn>=21.0.0

minio>=7.2.0
boto3>=1.34.0

# Databricks
databricks-sdk>=0.18.0
databricks-cli>=0.18.0
pyspark>=3.4.0

# Monitoring
evidently>=0.4.0
sentence-transformers>=2.2.0

python-dotenv>=1.0.0
tqdm>=4.66.0
requests>=2.31.0

pytest>=7.4.0
pytest-cov>=4.1.0

black>=23.12.0
isort>=5.13.0
flake8>=6.1.0
